requirements.txt
fastapi
uvicorn
torch
torchvision
ultralytics
opencv-python
pillow
pyyaml
requests
